Use the copy plugin instead of the plug plugin between contributions
This is necessary because there are times when using the plug plugin results in:

    snd1_pcm_hw_param_get_min: Assertion `!snd_interval_empty(i)' failed

Using the copy plugin puts the onus on Volumio plugins to set up rate/format
conversion if they need it. The Software Volume control therefore does a
conversion to S24_LE which should improve sound quality at low volume.

There is always a plug directly before the hardware stage. This is to allow
plugins not to worry about what parameters the hardware supports.