Ensure that plugins are notified of changes in the ALSA configuration
When a plugin changes the ALSA config (by requesting a rewrite) then we need to ensure that other plugins which are holding open an ALSA library instance know that they have to reload it. This is done by rebinding the alsa.outputdevice property, even though its value does not change from 'volumio'.